# backend/internal/httpapi/panelsettings/docs/swagger.yaml
basePath: /api
definitions:
...
  shared.ErrorResponse:
    properties:
      error:
        example: invalid input
        type: string
      message:
        example: invalid input
        type: string
    type: object
...
  users.OptionalInt:
    properties:
      set:
        type: boolean
      value:
        type: integer
    type: object
  users.OptionalInt64:
    properties:
      set:
        type: boolean
      value:
        format: int64
        type: integer
    type: object
  users.OptionalString:
    properties:
      set:
        type: boolean
      value:
        type: string
    type: object
...
  users.bulkAllExtendExpirationDateRequest:
    properties:
      extendDays:
        type: integer
    type: object
  users.bulkAllUpdateUsersRequest:
    properties:
      description:
        $ref: '#/definitions/users.OptionalString'
      email:
        $ref: '#/definitions/users.OptionalString'
      expireAt:
        type: string
      hwidDeviceLimit:
        $ref: '#/definitions/users.OptionalInt'
      status:
        type: string
      tag:
        $ref: '#/definitions/users.OptionalString'
      telegramId:
        $ref: '#/definitions/users.OptionalInt64'
      trafficLimitBytes:
        type: integer
      trafficLimitStrategy:
        type: string
    type: object
  users.bulkDeleteUsersRequest:
    properties:
      userIds:
        items:
          type: integer
        type: array
    type: object
  users.bulkExtendExpirationDateRequest:
    properties:
      extendDays:
        type: integer
      userIds:
        items:
          type: integer
        type: array
    type: object
  users.bulkRevokeUsersSubscriptionRequest:
    properties:
      userIds:
        items:
          type: integer
        type: array
    type: object
  users.bulkUpdateUsersFields:
    properties:
      description:
        $ref: '#/definitions/users.OptionalString'
      email:
        $ref: '#/definitions/users.OptionalString'
      expireAt:
        type: string
      externalSquadUuid:
        $ref: '#/definitions/users.OptionalString'
      hwidDeviceLimit:
        $ref: '#/definitions/users.OptionalInt'
      status:
        type: string
      tag:
        $ref: '#/definitions/users.OptionalString'
      telegramId:
        $ref: '#/definitions/users.OptionalInt64'
      trafficLimitBytes:
        type: integer
      trafficLimitStrategy:
        type: string
    type: object
  users.bulkUpdateUsersRequest:
    properties:
      fields:
        $ref: '#/definitions/users.bulkUpdateUsersFields'
      userIds:
        items:
          type: integer
        type: array
    type: object
  users.bulkUpdateUsersSquadsRequest:
    properties:
      activeInternalSquads:
        items:
          type: string
        type: array
      userIds:
        items:
          type: integer
        type: array
    type: object
...
info:
  contact: {}
  description: Exodus dashboard and node-management API.
  license:
    name: AGPL-3.0
  title: Exodus API
  version: 3.0.0
paths:
...
  /users/bulk/all/extend-expiration-date:
    post:
      consumes:
      - application/json
      description: Add days to expireAt for all users in the system
      parameters:
      - description: Extension days
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkAllExtendExpirationDateRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Extend expiration date for all users
      tags:
      - Users Bulk Actions Controller
  /users/bulk/all/reset-traffic:
    post:
      description: Reset used traffic to 0 for all users in the system
      responses:
        "202":
          description: Accepted
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Reset traffic for all users
      tags:
      - Users Bulk Actions Controller
  /users/bulk/all/update:
    post:
      consumes:
      - application/json
      description: Update fields for all users in the system
      parameters:
      - description: Fields to update across all users
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkAllUpdateUsersRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update all users
      tags:
      - Users Bulk Actions Controller
  /users/bulk/delete:
    post:
      consumes:
      - application/json
      description: Delete multiple users by their IDs
      parameters:
      - description: User IDs to delete
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkDeleteUsersRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk delete users by IDs
      tags:
      - Users Bulk Actions Controller
  /users/bulk/delete-by-status:
    post:
      consumes:
      - application/json
      description: Delete all users having a specific status
      parameters:
      - description: Status to delete
        in: body
        name: body
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk delete users by status
      tags:
      - Users Bulk Actions Controller
  /users/bulk/extend-expiration-date:
    post:
      consumes:
      - application/json
      description: Add days to expireAt for specified users
      parameters:
      - description: User IDs and extension days
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkExtendExpirationDateRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk extend users expiration date
      tags:
      - Users Bulk Actions Controller
  /users/bulk/reset-traffic:
    post:
      consumes:
      - application/json
      description: Reset used traffic to 0 for specified users
      parameters:
      - description: User IDs to reset
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkDeleteUsersRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk reset users traffic
      tags:
      - Users Bulk Actions Controller
  /users/bulk/revoke-subscription:
    post:
      consumes:
      - application/json
      description: Revoke subscriptions for multiple users
      parameters:
      - description: User IDs to revoke
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkRevokeUsersSubscriptionRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk revoke users subscription
      tags:
      - Users Bulk Actions Controller
  /users/bulk/update:
    post:
      consumes:
      - application/json
      description: Update fields for multiple users
      parameters:
      - description: User IDs and fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkUpdateUsersRequest'
      responses:
        "202":
          description: Accepted
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk update users
      tags:
      - Users Bulk Actions Controller
  /users/bulk/update-squads:
    post:
      consumes:
      - application/json
      description: Assign internal squads for multiple users
      parameters:
      - description: User IDs and internal squad UUIDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/users.bulkUpdateUsersSquadsRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk update users squads
      tags:
      - Users Bulk Actions Controller
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"